requirements.txt
# Core LangChain and Google Gemini support
langchain-google-genai
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
chromadb

# Data models and parsing
pydantic

# Language processing
nltk

# File loaders
PyPDFLoader

# Environment variable management
python-dotenv


rich
